+++ src/constants/TelemetryEvent.ts
@@ -22,6 +22,7 @@ export const TelemetryEvent = {
   uploadMedia: 'uploadMedia',
   refreshMedia: 'refreshMedia',
   deleteMedia: 'deleteMedia',
+  insertContentSnippet: 'insertContentSnippet',
   insertMediaToContent: 'insertMediaToContent',
   insertFileToContent: 'insertFileToContent',
   updateMediaMetadata: 'updateMediaMetadata',