+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - [#388](https://github.com/estruyf/vscode-front-matter/issues/388): New stop server action has been added to the panel
 - [#390](https://github.com/estruyf/vscode-front-matter/issues/390): Implement another JSON parser in order to be able to parse the `frontmatter.json` file better
 - [#394](https://github.com/estruyf/vscode-front-matter/issues/394): Ordering of snippet fields is based on their field definition
+- [#395](https://github.com/estruyf/vscode-front-matter/issues/395): Added support for custom snippet fields on media snippets
 - [#402](https://github.com/estruyf/vscode-front-matter/issues/402): Custom sorting of content now supports `number` fields
 
 ### ⚡️ Optimizations